pyproject.toml
[project]
name = "ai-exercise"
version = "0.1.0"
description = ""
authors = [{ name = "Matt Carey", email = "matt@stackone.com" }]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.4",
    "watchdog>=3.0.0",
    "chromadb>=0.5.18",
    "fastapi>=0.115.5",
    "langchain-text-splitters>=0.3.2",
    "pydantic-settings>=2.6.1",
    "pydantic-ai>=0.0.43",
]

[project.optional-dependencies]
dev = ["mypy>=1.2", "pytest>=7", "pytest-cov>=3", "ruff>=0.7.3"]
streamlit-demo = ["streamlit>=1.25.0", "watchdog>=3.0.0"]
evals = ["click>=8.1.8", "httpx>=0.28.1", "rich>=13.9.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


## ruff
[tool.ruff]
target-version = 'py311'

# See all rules at https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "W",   # pycodestyle
    "F",   # Pyflakes
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "D",   # pydocstyle
    "I",   # isort
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "UP",  # pyupgrade
    "PT",  # flake8-pytest-style
    "RUF", # Ruff-specific rules
]

ignore = [
    "D415", #annoying and unnecessary to need punctuation
    "D103", #I am lazy and I don't care about doc strings
    "W505",
]

[tool.ruff.lint.pydocstyle]
# Automatically disable rules that are incompatible with Google docstring convention
convention = "google"

## mypy

[tool.mypy]
python_version = "3.11"
# Enable all optional error checking flags, providing stricter type checking; see https://mypy.readthedocs.io/en/stable/getting_started.html#strict-mode-and-configuration
strict = true

# Type-check the interiors of functions without type annotations; if missing, mypy won't check function bodies without type hints, for instance those coming from third-party libraries
check_untyped_defs = true

# Make __init__.py file optional for package definitions; if missing, mypy requires __init__.py at packages roots, see https://mypy.readthedocs.io/en/stable/running_mypy.html#mapping-file-paths-to-modules
explicit_package_bases = true
